whisper_app/backend/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosmtplib==5.0.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
awscli==1.42.56
bcrypt==4.0.1
blinker==1.9.0
botocore==1.40.56
cachetools==6.2.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
cloudinary==1.44.1
colorama==0.4.6
cryptography==46.0.2
Deprecated==1.2.18
distro==1.9.0
dnspython==2.8.0
docutils==0.19
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.115.0
filelock==3.20.0
Flask==3.1.2
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
frozenlist==1.7.0
fsspec==2025.10.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-genai==1.55.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.75.1
grpcio-status==1.71.2
h11==0.16.0
http_ece==1.2.1
httpcore==1.0.9
httplib2==0.31.0
httptools==0.6.4
httpx==0.28.1
idna==3.10
ImageIO==2.37.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.0
jmespath==1.0.1
limits==5.6.0
llvmlite==0.45.1
loguru==0.7.3
MarkupSafe==3.0.2
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.6.4
networkx==3.5
numba==0.62.1
numpy==2.3.5
onesignal-sdk==2.0.0
openai==2.1.0
openai-whisper==20250625
packaging==25.0
passlib==1.7.4
pillow==11.3.0
propcache==0.4.0
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.10
py-vapid==1.9.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.2
pydantic-settings==2.11.0
pydantic_core==2.41.4
pydub==0.25.1
PyJWT==2.10.1
PyMySQL==1.1.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-http-client==3.3.7
python-jose==3.5.0
python-magic==0.4.27
python-multipart==0.0.20
pytz==2025.2
pywebpush==2.1.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
resend==2.19.0
rsa==4.7.2
s3transfer==0.14.0
sendgrid==6.12.5
setuptools==80.9.0
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.38.6
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.12.0
torch==2.9.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
uritemplate==4.2.0
urllib3==2.5.0
user-agents==2.2.0
uvicorn==0.37.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
win32_setctime==1.2.0
wrapt==1.17.3
yarl==1.21.0
